Update Firefox Kernel v98

+++ js/src/wasm/WasmTypeDef.h
@@ -829,7 +829,7 @@ class TypeIdDesc {
 
   bool isGlobal() const { return kind_ == TypeIdDescKind::Global; }
 
-  size_t immediate() const {
+  uint32_t immediate() const {
     MOZ_ASSERT(kind_ == TypeIdDescKind::Immediate);
     return bits_;
   }
